fix revisi module list summary
